# improve performance for string(...)