Before:

```
julia> @btime string('a') # what is even going on here...
  732.898 ns (4 allocations: 192 bytes)
"a"

julia> @btime string('a', "aa")
  168.540 ns (4 allocations: 192 bytes)
"aaa"

julia> @btime string("a", 'a', 3)
  467.898 ns (6 allocations: 288 bytes)
"aa3"

julia> @btime string('a', 3, "fdsfhjsdfjdlsfjldsjflsdjfkldsjflksdjflkdsjfldsjfk")
  537.153 ns (7 allocations: 368 bytes)
"a3fdsfhjsdfjdlsfjldsjflsdjfkldsjflksdjflkdsjfldsjfk"
```

After:

```
julia> @btime string('a')
  12.222 ns (1 allocation: 32 bytes)
"a"

julia> @btime string('a', "aa")
  37.108 ns (2 allocations: 64 bytes)
"aaa"

julia> @btime string("a", 'a', 3)
  293.483 ns (5 allocations: 272 bytes)
"aa3"

julia> @btime string('a', 3, "fdsfhjsdfjdlsfjldsjflsdjfkldsjflksdjflkdsjfldsjfk")
  313.825 ns (5 allocations: 320 bytes)
"a3fdsfhjsdfjdlsfjldsjflsdjfkldsjflksdjflkdsjfldsjfk"
```

Ref #28870