@KristofferC 2018-08-25
Why can't we bring back the previous optimized version for `Union{String, Char}`?

```
julia> function string2(a::Union{String,AbstractChar}...)
           n = 0
           for v in a
               if v isa Char
                   n += Base.codelen(v)
               else
                   n += sizeof(v)
               end
           end
           out = Base._string_n(n)
           offs = 1
           for v in a
               if v isa Char
                   x = bswap(reinterpret(UInt32, v))
                   for j in 1:Base.codelen(v)
                       unsafe_store!(pointer(out, offs), x % UInt8)
                       offs += 1
                       x >>= 8
                   end
               else
                   unsafe_copyto!(pointer(out,offs), pointer(v), sizeof(v))
                   offs += sizeof(v)
               end
           end
           return out
       end
string2 (generic function with 1 method)

julia> @btime string2('a', "fsf", 'a', "fdsf")
  42.563 ns (1 allocation: 32 bytes)
"afsfafdsf"

julia> @btime string('a', "fsf", 'a', "fdsf")
  259.211 ns (3 allocations: 176 bytes)
```