@bkamins 2018-08-25
Good point - we can . Though the signature should be `Union{String, SubString{String}, Char}` not `Union{String, AbstractChar}` as we should handle substrings and characters other than `Char` might have a different implementation.

As you are probably aware what you propose is not the "old" implementation of `Union{String, AbstarctChar}` which used `print` and would be slower, but it is a very nice piece of code combining all what we had discussed up to this point 👍.