@KristofferC 2018-08-25
Yeah, the optimization I was referring to bringing back was the one that used to look like: 

https://github.com/JuliaLang/julia/blob/7711781875620866da6a4f7dc545b355e5b1f24c/base/strings/string.jl#L350-L392

I pushed an update. I kept the previos modifications to `print_to_string` since the improvements there are still valid. Please review :)