@bkamins 2018-08-25
All looks good to me. The only fear I had is whether compiler efficiently handles the situation of 3 element union and `isa` tests and it looks that it does as it should. Thank you.